requirements.txt
google-cloud-firestore
python-dotenv
requests
functions-framework
